requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Frontend
streamlit==1.31.0

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM
google-generativeai==0.3.2

# Document Parsing
beautifulsoup4==4.12.3
pymupdf==1.23.21
markdown==3.5.2

# Selenium
selenium==4.17.2

# Utilities
python-dotenv==1.0.1
requests==2.31.0
